RestController gecontroleerd in Postman en werkt zoals het behoord
