--- src/main/resources/data.sql
@@ -1,1 +1,0 @@
-INSERT INTO televisions () VALUES ();
